FIX: Replace `version` integer metric with new `version_info` metric (#16)
The previous `git ref-list --count` metric does not work with shallow
clones of Discourse. With the latest Discourse base image, the number
returned will not be useful for comparisons.

This commit replaces the `version` metric with a `version_info` metric.

This always has a value of `1`. It has a `revision` label containing
the current discourse core commit hash, and a `version` label containing
the human-friendly `Discourse::VERSION::STRING`.