FIX: Ensure collector loads webrick from plugin gem directory (#71)
Discourse core recently dropped webrick as a dependency. That broke this plugin because the collector process was using core's webrick, even though the plugin also specified it as a dependency in plugin.rb.

This commit updates the collector to load webrick from the correct location using a similar strategy to the prometheus_exporter gem.