FEATURE: Add `database` label to `active_record_connections_count`
Why this change?

Prior to this change, the `active_record_connections_count` metric is
simply a sum of all the connections for all the connection pools.
However, this metric is not helpful in a multisite environment where
there are multiple connection pools and when we need to figure out which
connection pool is holding on to a large amount of connections.