FEATURE: Export metric for highest sequence usage
Plugin already exports `postgres_highest_sequence`, but the value cannot
be reliable used for alerts because it is not linked with the maximum
value of the column that uses it.

In other words, a sequence may be used with an `integer` or a `bigint`
column, which have very different maximum values. The highest
`last_value` reported may be cause of concern for an `integer` column
because it is close to the limit, but not a problem for a `bigint`
column.