FEATURE: Tweak `postgres_highest_sequence` only for int columns (#98)
Tweak `postgres_highest_sequence` metric to export the highest sequence
only for integer columns. These are the most probable columns to run out
of values.

There are some exceptions to this rule to include `bigint` columns that
may be referenced in `integer` columns. For example, `reviewables.id` is
a `bigint` column, but `reviewable_scores.reviewable_id` is an integer.