FEATURE: Tweak `postgres_highest_sequence` only for int columns (#120)
* FEATURE: Tweak `postgres_highest_sequence` only for int columns

This is a reimplementation of commit 1015d9e3 and a revert of commit
98b8d44b.

The original implementation was using the REPLACE function in the
filter, which was causing the query to be very slow. For some databases,
the old query ran in ~6 seconds, while this optimized one runs in ~30ms.

Unfortunately, it is still slower than the previous one that executed in
~2ms.

* DEV: Use a scheduled job to update stats in bg

This is a follow up to previous commit to ensure that metric collection
is not slow for multisites with hundreds of sites.