FIX: Fork collector demon in a safer way.
Previously we were using the following pattern to fork a child process

```
if @pid = fork
  write_pid_file
  return
end

<Do something in child process>
```

However, we noticed that if any error occurs in the child process (ie.
after the `if @pid = fork` line), the error propogates to the parent
process as well causing it to crash. Instead, we can pass `fork` a block
instead and that does not cause the error to propogate to the parent
process.