DEV: add tmp_dir_available_bytes metric
The Discourse application and plugins utilise the `/tmp` directory for
local storage during request and job processing. If the directory has no
space remaining, attempting to create files will fail.

Exposing the amount of space available in the directory is useful in
environments where this directory is limited in size.

See t/111786.