FEATURE: Add `success` label to `(sidekiq|scheduled)_job_count`
Before this commit, `scheduled_job_count` represents the total number of
scheduled jobs executed regardless of its execution status while
`sidekiq_job_count` represents the total number of sidekiq job executed
successfuly only.

In order to allow the error rate for a particular job, we need to be
able to differentiate between successes and failures. This commit
seeks to do so by adding a `success` label to the `sidekiq_job_count`
and `scheduled_job_count` counters.