FIX: Switch duration web metrics to histogram type
This is a bug fix that switches the `http_.*_duration_seconds"
metrics from a summary to a histogram type because summaries cannot be
aggregated and that makes it practically useless in a multi-instance
setup. As such, switching to a histogram allows us to better support
both single instance and multi-instance setups.

This is unfortunately a breaking change but we have decided to accept
the risk given that there is likely to be low usage of this plugin.