FEATURE: Introduce `http_requests_<type>_duration_seconds` metrics (#142)
This commit introduces a new set of
`http_requests_<type>_duration_seconds` web metrics of histogram type.
The way the metrics are observed is similar to the
`http_<type>_duration_seconds` web metrics but of summary type. The
reason we are introducing a new set of web duration metrics of histogram
type is because summary type cannot be aggregated which is important for
a multi-instance setup.

In the near future, we intend to drop the `http_<type>_duration_seconds`
web metrics of summary type but will keep it around for now for
backwards compatibility purpose.